This one was tough
I had a hunch that you'd need to work out the periodicity of each sequence.
Tried brute forcing but that seemed to be taking too long.

I inspected the cycle times manually and was curious what the prime
factorisations would be.

Haven't deeply thought it through but each cycle I have has the length of the
instruction sequence (281) as one prime factor and something else as the other
prime factor.

I got to the result mainly by guesswork that it would be some multiple of the
prime factors :shrug: